patch: Added Movie Service Backend Test + Fixed bugs uncovered by Tests (#178)

+++ src/main/java/com/example/CineMatch/dto/TvDto.java
@@ -12,16 +12,16 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TvDto {
-    public int id;
+    public Long id;
     public String name;
     public String original_name;
     public String overview;
     public String first_air_date;
     public String poster_path;
     public String backdrop_path;
-    public double popularity;
-    public double vote_average;
-    public int vote_count;
+    public Double popularity;
+    public Double vote_average;
+    public Double vote_count;
     public String original_language;
     public List<Integer> genre_ids;
 }
